Add 'default' arg to get_gen_int/bool, simpler alternative to default_gen_int
I think putting a default in default_gen_int and limit in gen_int_limits is
a bit too much for some unimportant or throwaway uses.

And I also made the limits optional, as they might not be static and need to be
enforced elsewhere. For example, to remember the selected menu item in an
editor.