WIP play accept sound when opening a menu. Add "open menu" arg to disable it.
Currently, "main menu" and "open menu" don't play AcceptSfx. "main menu"
definitely should:

Each of the special menus like Items etc play the accept sound when
they're opened, the menu itself doesn't play an accept sound when a menu item is
used!
The accept sound is played when the main menu is opened by the player
but that's again a special case, it doesn't happen when a menu is opened in any
other way.

A textbox that opens a menu should play AcceptSfx, just like
if it has a choice box.  When a menu is opened from a menu item, it should play
AcceptSfx.  An activated menu item with 'close when selected' should play
CancelSfx

When you activate a menu item, which sound effect ought to be is played depends
on what happens? If it runs a script, that script might or might not close the
menu