Auto-enable Bug #430 emulation bit in ubersetzung and older games
That is, games last edited in ubersetzung.
This had been the intention when I added the backcompat bit 12 years ago...

By a stroke of good luck, menuitem.bin was added on upgrade() just a few days
before the commit that permanently enabled the bug 430 fix, so we have a precise
rule for the turning on the fix, except...

Voxhumana stable-release games, but not contemporary wip-created games, ought to
get the bugfix rollback too, as the bug was renabled on that branch, but we
can't distinguish them from the wip ones.

I also restored the whatsnew note for the original fix, under werewaffle. Thanks
to Yuriy for helping to isolate bug 430.